requirements.txt
openai
modal
modal-client
bs4
openai-whisper
